wallbash enhancement

+++ Configs/.config/hypr/scripts/wbarstylegen.sh
@@ -10,11 +10,14 @@ modules_dir="$waybar_dir/modules"
 in_file="$waybar_dir/modules/style.css"
 out_file="$waybar_dir/style.css"
 src_file="$HOME/.config/hypr/themes/theme.conf"
+reload_flag=0
 
 if [ "$EnableWallDcol" -eq 1 ] ; then
-    export cur_theme="Wall-Dcol"
-else
-    export cur_theme=`gsettings get org.gnome.desktop.interface gtk-theme | sed "s/'//g"`
+    ln -fs $waybar_dir/themes/Wall-Dcol.css $waybar_dir/themes/theme.css
+    reload_flag=1
+elif [ $(readlink $waybar_dir/themes/theme.css) != "$waybar_dir/themes/${gtkTheme}.css" ] ; then
+    ln -fs $waybar_dir/themes/${gtkTheme}.css $waybar_dir/themes/theme.css
+    reload_flag=1
 fi
 
 
@@ -67,8 +70,8 @@ fi
 
 # restart waybar
 
-killall waybar
-waybar > /dev/null 2>&1 &
-# killall -SIGUSR2 waybar
-
-
+if [ $reload_flag -eq 1 ] ; then
+    killall waybar
+    waybar > /dev/null 2>&1 &
+    # killall -SIGUSR2 waybar
+fi
